+++ src/ccb/image/Image.hpp
@@ -212,17 +212,17 @@ namespace ccb { namespace image
             return this->data.size();
         }
 
-        uint32_t GetWidth() const
+        size_t GetWidth() const
         {
             return this->width;
         }
 
-        uint32_t GetHeight() const
+        size_t GetHeight() const
         {
             return this->height;
         }
 
-        uint32_t GetStride() const
+        size_t GetStride() const
         {
             return this->stride;
         }
@@ -312,7 +312,7 @@ namespace ccb { namespace image
                 throw std::logic_error("Unregistered pixel type");
             }
 
-            return Image(std::move(data), width, height, stride, typeCode);
+            return Image(std::move(data), width, height, stride, static_cast<unsigned>(typeCode));
         }
     };
 } }
